+++ src/sentry_batcher.h
@@ -2,6 +2,7 @@
 #define SENTRY_BATCHER_H_INCLUDED
 
 #include "sentry_boot.h"
+#include "sentry_client_report.h"
 #include "sentry_database.h"
 #include "sentry_envelope.h"
 #include "sentry_sync.h"
@@ -44,6 +45,7 @@ typedef struct {
     sentry_waitable_flag_t request_flush; // level-triggered flush flag
     sentry_threadid_t batching_thread; // the batching thread
     sentry_batch_func_t batch_func; // function to add items to envelope
+    sentry_data_category_t data_category; // for client report discard tracking
     sentry_dsn_t *dsn;
     sentry_transport_t *transport;
     sentry_run_t *run;
@@ -57,7 +59,8 @@ typedef struct {
 
 #define SENTRY_BATCHER_REF_INIT { NULL, 0 }
 
-sentry_batcher_t *sentry__batcher_new(sentry_batch_func_t batch_func);
+sentry_batcher_t *sentry__batcher_new(
+    sentry_batch_func_t batch_func, sentry_data_category_t data_category);
 
 /**
  * Acquires a reference to the batcher behind `ref`, atomically incrementing