+++ src/sentry_core.c
@@ -5,6 +5,7 @@
 
 #include "sentry_attachment.h"
 #include "sentry_backend.h"
+#include "sentry_client_report.h"
 #include "sentry_core.h"
 #include "sentry_database.h"
 #include "sentry_envelope.h"
@@ -623,6 +624,8 @@ sentry__capture_event(sentry_value_t event, sentry_scope_t *local_scope)
             bool should_skip = !sentry__roll_dice(options->sample_rate);
             if (should_skip) {
                 SENTRY_INFO("throwing away event due to sample rate");
+                sentry__client_report_discard(SENTRY_DISCARD_REASON_SAMPLE_RATE,
+                    SENTRY_DATA_CATEGORY_ERROR, 1);
                 sentry_envelope_free(envelope);
             } else {
                 sentry__capture_envelope(options->transport, envelope);
@@ -716,6 +719,8 @@ sentry__prepare_event(const sentry_options_t *options, sentry_value_t event,
             = options->before_send_func(event, NULL, options->before_send_data);
         if (sentry_value_is_null(event)) {
             SENTRY_DEBUG("event was discarded by the `before_send` hook");
+            sentry__client_report_discard(SENTRY_DISCARD_REASON_BEFORE_SEND,
+                SENTRY_DATA_CATEGORY_ERROR, 1);
             return NULL;
         }
     }
@@ -767,6 +772,8 @@ sentry__prepare_transaction(const sentry_options_t *options,
         if (sentry_value_is_null(transaction)) {
             SENTRY_DEBUG(
                 "transaction was discarded by the `before_transaction` hook");
+            sentry__client_report_discard(SENTRY_DISCARD_REASON_BEFORE_SEND,
+                SENTRY_DATA_CATEGORY_TRANSACTION, 1);
             return NULL;
         }
     }